+++ Scrumr.Client/Extensions.cs
@@ -14,12 +14,12 @@ namespace Scrumr.Client
 {
     static class Extensions
     {
-        public static T Get<T>(this IEnumerable<T> target, long id) where T : Entity
+        public static T Get<T>(this IEnumerable<T> target, int id) where T : Entity
         {
             return target.SingleOrDefault(x => x.ID == id);
         }
 
-        public static bool Has<T>(this IEnumerable<T> target, long id) where T : Entity
+        public static bool Has<T>(this IEnumerable<T> target, int id) where T : Entity
         {
             return target.Get(id) != null;
         }
